+++ src/components/telemetry_monitor/test/protocol_handler_metric_test.cc
@@ -103,9 +103,9 @@ TEST(ProtocolHandlerMetricTest, GetJsonMetricWithGrabResources) {
   EXPECT_EQ(5, jvalue[strings::message_id].asInt64());
   EXPECT_EQ(2, jvalue[strings::connection_key].asInt());
 
-  EXPECT_EQ(resources->stime, jvalue[strings::stime].asInt());
-  EXPECT_EQ(resources->utime, jvalue[strings::utime].asInt());
-  EXPECT_EQ(resources->memory, jvalue[strings::memory].asInt());
+  EXPECT_EQ(resources->stime, jvalue[strings::stime].asInt64());
+  EXPECT_EQ(resources->utime, jvalue[strings::utime].asInt64());
+  EXPECT_EQ(resources->memory, jvalue[strings::memory].asInt64());
 }
 
 }  // namespace telemetry_monitor_test
